fixed a bug tha caused an infinite loop
When there are two or more closing brackets without an opening one
between them, the compiler would generate an infinite loop. The downside
is that the decompiled C# code is less readable. I should add a test
case for this and treat it as a special case to generate cleaner IL.